requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
azure-core==1.33.0
azure-identity==1.14.1
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.25.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
docx==0.2.4
docx2txt==0.9
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
extra-streamlit-components==0.1.71
faiss-cpu==1.11.0
filetype==1.2.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.10
importlib_metadata==7.2.1
isodate==0.7.2
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.1.16
langchain-community==0.0.38
langchain-core==0.1.53
langchain-openai==0.1.7
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.147
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msal==1.24.0b2
msal-extensions==1.0.0
msal-streamlit-authentication==1.0.9
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.37.0
nltk==3.9.1
numpy==1.26.4
openai==1.31.2
openpyxl==3.1.2
orjson==3.10.16
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.3
pgvector==0.4.1
pillow==10.4.0
plum-dispatch==1.7.4
portalocker==2.10.1
propcache==0.3.1
protobuf==4.25.7
psycopg2-binary
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.0
pypdf==5.4.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2025.2.18
PyYAML==6.0.2
random-password-generator==2.2.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
selenium==4.31.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spire-doc==13.3.8
SQLAlchemy==2.0.40
streamlit==1.45.0
streamlit-authenticator==0.1.5
streamlit-modal==0.1.2
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.11.0
urllib3==2.4.0
validators==0.34.0
watchdog==6.0.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
supabase
SpeechRecognition 
pocketsphinx
streamlit-audiorec
matplotlib
seaborn

